requirements.txt
pandas==2.1.4
numpy==1.26.2
yfinance>=0.2.36
requests==2.31.0
python-dotenv==1.0.0
pandas-ta==0.3.14b
pyyaml==6.0.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
